Safari-609.1.11.1/LayoutTests/fast/dom/SelectorAPI/tag-case-insensitive-expected.txt

PASS document.querySelector('SPAN') is document.getElementById('upper1')
PASS document.getElementById('upper1').webkitMatchesSelector('SPAN') is true
PASS document.querySelector('SPAN > img') is document.getElementById('img1')
PASS document.getElementById('img1').webkitMatchesSelector('SPAN > img') is true
PASS document.querySelector('SPAN img') is document.getElementById('img1')
PASS document.getElementById('img1').webkitMatchesSelector('SPAN img') is true
PASS document.querySelector('div ul li SPAN') is document.getElementById('upper1')
PASS document.getElementById('upper1').webkitMatchesSelector('div ul li SPAN') is true
PASS document.querySelector('div ul li textarea + SPAN') is document.getElementById('upper1')
PASS document.getElementById('upper1').webkitMatchesSelector('div ul li textarea + SPAN') is true
PASS document.querySelector('div ul li span') is document.getElementById('upper1')
PASS document.getElementById('upper1').webkitMatchesSelector('div ul li span') is true
PASS document.querySelector('div ul li spAn') is document.getElementById('upper1')
PASS document.getElementById('upper1').webkitMatchesSelector('div ul li spAn') is true
PASS document.querySelector('DIV ul li span') is document.getElementById('upper1')
PASS document.getElementById('upper1').webkitMatchesSelector('DIV ul li span') is true
PASS document.querySelector('div UL li span') is document.getElementById('upper1')
PASS document.getElementById('upper1').webkitMatchesSelector('div UL li span') is true
PASS content.querySelector('span1') is null
PASS content.querySelector('span#UPPER1') is null
PASS document.querySelector('unknownTag') is document.getElementById('unknownTag1')
PASS document.getElementById('unknownTag1').webkitMatchesSelector('unknownTag') is true
PASS document.querySelector('div unknownTag') is document.getElementById('unknownTag1')
PASS document.getElementById('unknownTag1').webkitMatchesSelector('div unknownTag') is true
PASS document.querySelector('ul + unknownTag') is document.getElementById('unknownTag1')
PASS document.getElementById('unknownTag1').webkitMatchesSelector('ul + unknownTag') is true
PASS document.querySelector('DIV unknownTag') is document.getElementById('unknownTag1')
PASS document.getElementById('unknownTag1').webkitMatchesSelector('DIV unknownTag') is true
PASS document.querySelector('UL + unknownTag') is document.getElementById('unknownTag1')
PASS document.getElementById('unknownTag1').webkitMatchesSelector('UL + unknownTag') is true
PASS document.querySelector('unknowntag') is document.getElementById('unknownTag1')
PASS document.getElementById('unknownTag1').webkitMatchesSelector('unknowntag') is true
PASS document.querySelector('Unknowntag') is document.getElementById('unknownTag1')
PASS document.getElementById('unknownTag1').webkitMatchesSelector('Unknowntag') is true
PASS document.querySelector('div unknowntag') is document.getElementById('unknownTag1')
PASS document.getElementById('unknownTag1').webkitMatchesSelector('div unknowntag') is true
PASS document.querySelector('ul') is document.getElementById('ul1')
PASS document.getElementById('ul1').webkitMatchesSelector('ul') is true
PASS document.querySelector('UL') is document.getElementById('ul1')
PASS document.getElementById('ul1').webkitMatchesSelector('UL') is true
PASS content.querySelector('unknowntag1') is null
PASS content.querySelector('ol') is null
PASS content.querySelector('OL') is null
PASS successfullyParsed is true

TEST COMPLETE
